Remove tab layout from watch list
