+++ app/src/main/java/com/example/traders/watchlist/WatchListFragment.kt
@@ -6,22 +6,19 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.viewModels
 import androidx.lifecycle.lifecycleScope
-import androidx.navigation.NavController
-import androidx.viewpager2.widget.ViewPager2
 import com.example.traders.BaseFragment
-import com.example.traders.R
 import com.example.traders.databinding.FragmentWatchListBinding
-import com.example.traders.watchlist.adapters.ViewPagerAdapter
-import com.google.android.material.tabs.TabLayout
-import com.google.android.material.tabs.TabLayoutMediator
+import com.example.traders.watchlist.adapters.SingleCryptoListener
+import com.example.traders.watchlist.adapters.WatchListAdapter
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.launch
-import javax.inject.Inject
-import javax.inject.Provider
 
 @AndroidEntryPoint
 class WatchListFragment : BaseFragment() {
+
+    private val viewModel: WatchListViewModel by viewModels()
+    private var adapter: WatchListAdapter? = null
 
     override fun onCreateView(
         inflater: LayoutInflater,
@@ -29,20 +26,51 @@ class WatchListFragment : BaseFragment() {
         savedInstanceState: Bundle?
     ): View {
         val binding = FragmentWatchListBinding.inflate(layoutInflater, container, false)
-        setUpTabs(binding.cryptoPageViewer, binding.cryptoTab)
+        with(viewModel) {
+            isLoading.observe(viewLifecycleOwner) {
+                binding.handleLoader(it)
+            }
+            lifecycleScope.launch {
+                state.collect { state ->
+                    binding.pullToRefresh.isRefreshing = state.isRefreshing
+                    adapter?.submitList(state.binanceCryptoData)
+                }
+            }
+        }
+
+        binding.setPullToRefreshListener()
+        binding.setListAdapter()
 
         return binding.root
     }
 
-    private fun setUpTabs(viewPager: ViewPager2, tabLayout: TabLayout) {
-        val viewPagerAdapter = ViewPagerAdapter(requireActivity().supportFragmentManager, lifecycle)
-        viewPager.adapter = viewPagerAdapter
-        TabLayoutMediator(tabLayout, viewPager) { tab, position ->
-            when (position) {
-                0 -> tab.text = getString(R.string.watchlist_tab_all_crypto_label)
-                1 -> tab.text = getString(R.string.watchlist_tab_favourites_label)
-                2 -> tab.text = getString(R.string.watchlist_tab_new_crypto_label)
+    private fun FragmentWatchListBinding.setPullToRefreshListener() {
+        pullToRefresh.setOnRefreshListener {
+            viewModel.getCryptoOnRefresh()
+        }
+    }
+
+    private fun FragmentWatchListBinding.setListAdapter() {
+        adapter = WatchListAdapter(SingleCryptoListener { slug, symbol ->
+            if (symbol != null) {
+                val direction = WatchListFragmentDirections
+                    .actionWatchListFragmentToCryptoItem(slug, symbol)
+                navController.navigate(direction)
             }
-        }.attach()
+        })
+        itemsList.adapter = adapter
+    }
+
+    private fun FragmentWatchListBinding.handleLoader(isLoading: Boolean) {
+        if (isLoading) {
+            allCryptoLoader.visibility = View.VISIBLE
+        } else {
+            allCryptoLoader.visibility = View.GONE
+        }
+    }
+
+    override fun onDestroy() {
+        super.onDestroy()
+        adapter = null
     }
 }
