+++ app/src/main/res/values/strings.xml
@@ -102,6 +102,7 @@
     <string name="empty_list_message">Deposit USD to begin trading</string>
     <string name="empty_transaction_list_message">No transactions yet</string>
     <string name="transaction_history">Transaction history</string>
+    <string name="watch_list">Watch list</string>
 
 </resources>
